+++ packages/ocean-core/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](https://github.com/ocean-ds/ocean-web/compare/v1.15.0...v1.16.0) (2021-08-16)
+
+### Features
+
+- add alert component ([#768](https://github.com/ocean-ds/ocean-web/issues/768)) ([cc924bb](https://github.com/ocean-ds/ocean-web/commit/cc924bb81f9b6de18f390ae367e1e1070181eea5))
+
 # [1.15.0](https://github.com/ocean-ds/ocean-web/compare/v1.14.0...v1.15.0) (2021-08-13)
 
 ### Features